A fast and wild refactor... so many changes...
Biggest thing that changed was deciding to use Option types in place
of C++-style designated 'invalid elements'. Those things are still there
for some reason...

Other stuff that changed was basically all iteration interfaces are gone
and most of the api is starting to be defined by traits to help reduce
effort for supporting a million very similar little operations.

Most tests will still be failing, tons of the kernel impl is
'unimplemented!', and I'm trying hard to pull on the reigns here a lot
to concentrate on the fundamentals before moving towards the more
ambitious ideas that I had.